+++ flowman-core/src/main/scala/com/dimajix/flowman/jdbc/SqlStatements.scala
@@ -78,6 +78,8 @@ abstract class SqlStatements {
     def dropIndex(table: TableIdentifier, indexName: String) : String
     def createIndex(table: TableIdentifier, index:TableIndex) : String
 
+    def dropConstraint(table: TableIdentifier, constraintName: String): String
+
     def merge(targetTable: TableIdentifier, targetAlias:String, targetSchema:Option[StructType], sourceAlias:String, sourceSchema:StructType, condition:Column, clauses:Seq[MergeClause]) : String
     def merge(targetTable: TableIdentifier, targetAlias:String, targetSchema:Option[StructType], sourceTable: TableIdentifier, sourceAlias:String, sourceSchema:StructType,  condition:Column, clauses:Seq[MergeClause]) : String
 }